complete/requirement.txt
asgiref==3.1.2
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.3.3
Automat==0.7.0
cachetools==3.1.0
channels==2.2.0
constantly==15.1.0
daphne==2.3.0
Django==2.2.1
django-crispy-forms==1.7.2
django-filter==2.1.0
djangorestframework==3.9.4
google-api-python-client==1.7.8
google-auth==1.6.3
google-auth-httplib2==0.0.3
httplib2==0.12.3
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
Markdown==3.1
oauth2client==4.1.3
Pillow==6.0.0
pkg-resources==0.0.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
PyHamcrest==1.9.0
pytz==2019.1
rsa==4.0
six==1.12.0
sqlparse==0.3.0
Twisted==19.2.0
txaio==18.8.1
uritemplate==3.0.0
zope.interface==4.6.0
